Add config_update module to base-modules.
https://www.drupal.org/project/config_update

From the description:
The main use case is: You update a module, and it has either changed
default config that it provides, or added new config items. You want
to be able to import the new items, view the differences between the
site config and the changed config, and possibly "revert" (or it may
be an update) to the newly-provided default config.